+++ tests/test_main.c
@@ -29,5 +29,6 @@ int main(void)
     suite_mutex();
     suite_semaphore();
     suite_atomic_queue();
+    suite_thread_pool();
     return UNITY_END();
 }